worker/requirements.txt
boto3==1.11.0
botocore==1.14.0
docutils==0.15.2
jmespath==0.9.4
pandas==0.24.2
pymongo==3.10.1
python-dateutil==2.8.1
pytz==2019.3
redis==3.3.11
s3transfer==0.3.0
six==1.13.0
urllib3==1.25.7
